+++ zallet-core/src/components/json_rpc/methods/get_operation.rs
@@ -5,9 +5,7 @@ use jsonrpsee::core::RpcResult;
 use schemars::JsonSchema;
 use serde::Serialize;
 
-use crate::components::json_rpc::asyncop::{
-    AsyncOperation, OperationId, OperationState, OperationStatus,
-};
+use crate::components::json_rpc::asyncop::{AsyncOperation, OperationId, OperationStatus};
 
 /// Response to a `z_getoperationstatus` or `z_getoperationresult` RPC request.
 pub(crate) type Response = RpcResult<ResultType>;
@@ -45,10 +43,7 @@ pub(crate) async fn result(
     let mut remove = HashSet::new();
 
     for op in filtered(async_ops, filter) {
-        if matches!(
-            op.state().await,
-            OperationState::Success | OperationState::Failed | OperationState::Cancelled
-        ) {
+        if op.state().await.is_finished() {
             ret.push(op.to_status().await);
             remove.insert(op.operation_id().clone());
         }